Custom __deepcopy__ for LazyLoader
When running JJB under py3 (py3.6) it started to fail with exception

  TypeError: cannot serialize '_io.BufferedReader' object

This is happens when doing deepcopy of a LazyLoader object which
contains an instance of BufferedReader object.
Under py2 this works but under py3 it is not.

This commit adds custom __deepcopy__ method to LazyLoader that
avoids deepcopy of a self._loader.

Change-Id: I00844fc06ebf84538d557c33f1179a26cb18202f
Story: 2007227