Implement disable discover_pr_origin
In my opinion the best choise for default value of discover_pr_origion
is not specified(None value) like it is done for bitbucket scm.
Currently there is no way to disable this option for github scm.
Added false value as one of valid value for this option, like is done
for discover-pr-fork-strategy. In fact any of yaml specification false
implementation should work.

Change-Id: Ibc26d6b5c39b5455160ba5c84be99575b79fc102